Move state check to func. Tidy files and functions.
